Extract upsell styles to per component css modules
