+++ frontend/src/metabase/admin/upsells/components/UpsellBanner.module.css
@@ -1,0 +1,17 @@
+@import "./Upsells.base.module.css";
+
+.UpsellBannerComponent {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 1rem;
+  padding: 1.25rem 2rem;
+  overflow: hidden;
+  border: 1px solid var(--mb-color-border);
+  border-radius: 0.5rem;
+  width: 100%;
+
+  &.Large {
+    align-items: flex-start;
+  }
+}
