+++ frontend/src/metabase/admin/upsells/components/UpsellBigCard.module.css
@@ -1,0 +1,14 @@
+.UpsellBigCardComponent {
+  border: 1px solid var(--mb-color-border);
+  border-radius: 0.5rem;
+  display: flex;
+  align-items: center;
+  flex-wrap: nowrap;
+  max-width: fit-content;
+  overflow: hidden;
+
+  img {
+    aspect-ratio: 4/3;
+    max-height: 21.875rem;
+  }
+}
