+++ frontend/src/metabase/admin/upsells/components/UpsellGem.module.css
@@ -1,0 +1,6 @@
+@import "./Upsells.base.module.css";
+
+.UpsellGem {
+  color: var(--mb-color-upsell-gem);
+  flex-shrink: 0;
+}
