+++ frontend/src/metabase/admin/upsells/components/UpsellPill.module.css
@@ -1,0 +1,23 @@
+@import "./Upsells.base.module.css";
+
+.UpsellPillComponent {
+  display: inline-flex;
+  gap: 0.5rem;
+  align-items: center;
+  justify-content: center;
+  flex-grow: 0;
+  font-weight: bold;
+  font-size: 0.75rem;
+  text-decoration: none;
+  padding: 0.25rem 0.75rem;
+  border-radius: 2rem;
+  border: 1px solid var(--mb-color-upsell-secondary);
+  color: var(--mb-color-upsell-primary);
+  cursor: pointer;
+
+  &:hover {
+    background-color: var(--mb-color-upsell-primary);
+    color: var(--mb-color-text-white);
+    border: 1px solid var(--mb-color-upsell-primary);
+  }
+}
