+++ frontend/src/metabase/admin/upsells/components/Upsells.module.css
@@ -1,86 +1,8 @@
 @import "./Upsells.base.module.css";
 
-.UpsellGem {
-  color: var(--mb-color-upsell-gem);
-  flex-shrink: 0;
-}
-
-.UpsellBannerComponent {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 1rem;
-  padding: 1.25rem 2rem;
-  overflow: hidden;
-  border: 1px solid var(--mb-color-border);
-  border-radius: 0.5rem;
-  width: 100%;
-
-  &.Large {
-    align-items: flex-start;
-  }
-}
-
 .SecondaryCTALink {
   display: inline;
   color: var(--mb-color-upsell-primary);
   font-weight: 700;
   margin-left: 0.25rem;
-}
-
-.UpsellBigCardComponent {
-  border: 1px solid var(--mb-color-border);
-  border-radius: 0.5rem;
-  display: flex;
-  align-items: center;
-  flex-wrap: nowrap;
-  max-width: fit-content;
-  overflow: hidden;
-
-  img {
-    aspect-ratio: 4/3;
-    max-height: 21.875rem;
-  }
-}
-
-.UpsellCardComponent {
-  border-radius: 0.5rem;
-  overflow: hidden;
-  border: 1px solid var(--mb-color-upsell-secondary);
-  background-color: var(--mb-color-bg-white);
-
-  &.Large {
-    .MainStack {
-      padding-top: 2rem;
-      padding-bottom: 3rem;
-    }
-  }
-
-  &.Default {
-    .Title {
-      font-size: 0.875rem;
-    }
-  }
-}
-
-.UpsellPillComponent {
-  display: inline-flex;
-  gap: 0.5rem;
-  align-items: center;
-  justify-content: center;
-  flex-grow: 0;
-  font-weight: bold;
-  font-size: 0.75rem;
-  text-decoration: none;
-  padding: 0.25rem 0.75rem;
-  border-radius: 2rem;
-  border: 1px solid var(--mb-color-upsell-secondary);
-  color: var(--mb-color-upsell-primary);
-  cursor: pointer;
-
-  &:hover {
-    background-color: var(--mb-color-upsell-primary);
-    color: var(--mb-color-text-white);
-    border: 1px solid var(--mb-color-upsell-primary);
-  }
 }
